modified based on the review

+++ problem/ch-08/07.ml
@@ -3,8 +3,8 @@ type ekikan_t = {
     kiten   : string;   (* 起点の駅名 *)
     shuten  : string;   (* 終点の駅名 *)
     keiyu   : string;   (* 経由する駅名 *)
-    kyori   : int;      (* 2駅間の距離 [km] *)
+    kyori   : float;    (* 2駅間の距離 [km] *)
     jikan   : int;      (* 所要時間 [分] *)
 }
 
-let shibuya_shinjuku = {kiten = "渋谷"; shuten = "新宿"; keiyu = "原宿"; kyori = 5; jikan = 5}
+let shibuya_shinjuku = {kiten = "渋谷"; shuten = "新宿"; keiyu = "原宿"; kyori = 5.; jikan = 5}
